# api/knowledge-base-openapi.yaml
openapi: 3.0.3
info:
  title: rag-hub API
  version: 1.0.0
  description: API specification for the phase-1 rag-hub document intelligence platform.
servers:
  - url: https://kb.example.com
    description: Production
  - url: http://test-kb.example.com
    description: Test
tags:
  - name: Auth
  - name: Documents
  - name: Tasks
  - name: Search
  - name: QA
  - name: Permissions
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain a Bearer JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/documents/upload:
    post:
      tags: [Documents]
      summary: Upload a single document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                biz_domain:
                  type: string
                department:
                  type: string
                security_level:
                  type: string
                  enum: [public, internal, confidential]
                source_system:
                  type: string
                owner:
                  type: string
                permission_tags:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/documents/batch-import:
    post:
      tags: [Documents]
      summary: Import documents in batch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchImportRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchImportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/documents/{documentId}/reparse:
    post:
      tags: [Documents]
      summary: Reparse current version of document
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReparseRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get task status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/documents:
    get:
      tags: [Documents]
      summary: Query documents
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: biz_domain
          schema:
            type: string
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page_no
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
  /api/documents/{documentId}:
    get:
      tags: [Documents]
      summary: Get document detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/documents/{documentId}/chunks:
    get:
      tags: [Documents]
      summary: Get chunks by document
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - in: query
          name: version_id
          schema:
            type: string
            format: uuid
        - in: query
          name: page_no
          schema:
            type: integer
        - in: query
          name: chunk_type
          schema:
            type: string
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
        - in: query
          name: page_no_num
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkListResponse'
  /api/search/query:
    post:
      tags: [Search]
      summary: Search only
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /api/qa/query:
    post:
      tags: [QA]
      summary: Retrieval augmented QA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QARequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QAResponse'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /api/query-logs/{logId}:
    get:
      tags: [QA]
      summary: Get query log detail
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: logId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryLogResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/documents/{documentId}/versions/{versionId}/activate:
    post:
      tags: [Documents]
      summary: Activate a document version
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DocumentId'
        - in: path
          name: versionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateVersionRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
  /api/permissions/bind:
    post:
      tags: [Permissions]
      summary: Bind permission policies to a resource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionBindRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentId:
      in: path
      name: documentId
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      in: path
      name: taskId
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: Gateway Timeout
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: string
          example: KB-00000
        message:
          type: string
          example: success
        trace_id:
          type: string
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: KB-40001
        message:
          type: string
        trace_id:
          type: string
        data:
          nullable: true
    UploadResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                document_id:
                  type: string
                  format: uuid
                version_id:
                  type: string
                  format: uuid
                task_id:
                  type: string
                  format: uuid
                status:
                  type: string
    BatchImportRequest:
      type: object
      required: [source_type, source_uri]
      properties:
        source_type:
          type: string
          enum: [directory, file_list]
        source_uri:
          type: string
        biz_domain:
          type: string
        department:
          type: string
        security_level:
          type: string
          enum: [public, internal, confidential]
    BatchImportResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                batch_id:
                  type: string
                task_count:
                  type: integer
                accepted_count:
                  type: integer
    ReparseRequest:
      type: object
      properties:
        force_reindex:
          type: boolean
          default: false
        reason:
          type: string
    TaskResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TaskData'
    TaskData:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        task_type:
          type: string
        status:
          type: string
        step:
          type: string
        retry_count:
          type: integer
        error_message:
          type: string
          nullable: true
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
    DocumentListResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total:
                  type: integer
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/DocumentItem'
    DocumentItem:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        title:
          type: string
        biz_domain:
          type: string
        department:
          type: string
        status:
          type: string
        current_version_id:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
    DocumentDetailResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/DocumentDetail'
    DocumentDetail:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        doc_code:
          type: string
        title:
          type: string
        source_type:
          type: string
        source_uri:
          type: string
        biz_domain:
          type: string
        department:
          type: string
        security_level:
          type: string
        status:
          type: string
        current_version:
          $ref: '#/components/schemas/DocumentVersion'
    DocumentVersion:
      type: object
      properties:
        version_id:
          type: string
          format: uuid
        version_no:
          type: string
        parse_status:
          type: string
        index_status:
          type: string
        effective_from:
          type: string
          format: date-time
        effective_to:
          type: string
          format: date-time
        is_current:
          type: boolean
    ChunkListResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total:
                  type: integer
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ChunkItem'
    ChunkItem:
      type: object
      properties:
        chunk_id:
          type: string
          format: uuid
        chunk_no:
          type: integer
        chunk_type:
          type: string
        title_path:
          type: string
        locator:
          type: string
        page_no:
          type: integer
        content_text:
          type: string
        content_summary:
          type: string
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        top_k:
          type: integer
          default: 10
          maximum: 20
        filters:
          $ref: '#/components/schemas/SearchFilters'
    SearchFilters:
      type: object
      properties:
        biz_domain:
          type: string
        department:
          type: string
        security_level:
          type: string
    SearchResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total:
                  type: integer
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/SearchItem'
    SearchItem:
      type: object
      properties:
        chunk_id:
          type: string
        document_id:
          type: string
        document_title:
          type: string
        title_path:
          type: string
        locator:
          type: string
        score:
          type: number
          format: float
        snippet:
          type: string
    QARequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        top_k:
          type: integer
          default: 6
          maximum: 20
        return_citations:
          type: boolean
          default: true
        filters:
          $ref: '#/components/schemas/SearchFilters'
        session_id:
          type: string
    QAResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                answer:
                  type: string
                confidence:
                  type: number
                  format: float
                citations:
                  type: array
                  items:
                    $ref: '#/components/schemas/CitationItem'
                retrieved_count:
                  type: integer
                used_chunk_count:
                  type: integer
                session_id:
                  type: string
    CitationItem:
      type: object
      properties:
        document_id:
          type: string
        document_title:
          type: string
        title_path:
          type: string
        locator:
          type: string
        snippet:
          type: string
    QueryLogResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                log_id:
                  type: string
                query_text:
                  type: string
                rewritten_query:
                  type: string
                answer_text:
                  type: string
                latency_ms:
                  type: integer
                citations:
                  type: array
                  items:
                    $ref: '#/components/schemas/CitationItem'
    ActivateVersionRequest:
      type: object
      properties:
        effective_from:
          type: string
          format: date-time
        remark:
          type: string
    PermissionBindRequest:
      type: object
      required: [resource_type, resource_id, policies]
      properties:
        resource_type:
          type: string
          enum: [document, chunk]
        resource_id:
          type: string
          format: uuid
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PermissionPolicy'
    PermissionPolicy:
      type: object
      required: [subject_type, subject_value, effect]
      properties:
        subject_type:
          type: string
          enum: [user, role, department]
        subject_value:
          type: string
        effect:
          type: string
          enum: [allow, deny]

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                tokenType:
                  type: string
                  example: Bearer
                accessToken:
                  type: string
                expiresInSeconds:
                  type: integer
                user:
                  type: object
                  properties:
                    userId:
                      type: string
                      format: uuid
                    username:
                      type: string
                    displayName:
                      type: string
                    roleCode:
                      type: string